pyproject.toml
[tool.poetry]
name = "orm-hw"
version = "0.1.0"
description = ""
authors = ["glebk <glebkot57@gmail.com>"]
readme = "README.md"
packages = [{include = "orm_hw"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
automium-web = "^0.1.1"
django-autoslug = "^1.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
